Also output gateway in debug message
The libipsec binding was not showing the gateway in the
debug message announcing route registration, however
it did try to register a route with a gateway.

This is confusing since the pfkey plugin does dump the
gateway ip when registering route.

This commit makes sure libipsec also dumps the gateway ip.